phy: renesas: rcar-gen3-usb2: No need to request IRQ for non-OTG
As stated in the comment found on top of rcar_gen3_phy_usb2_work,
we should not be registering the IRQ if !otg_channel || !uses_otg_pins.
On top of that, we should also make sure that extcon has been
allocated before requesting the irq as rcar_gen3_phy_usb2_work
uses it, hence the patch.

Fixes: 2b38543c8db1 ("phy: rcar-gen3-usb2: add extcon support")
Fixes: 73801b90a38f ("phy: renesas: rcar-gen3-usb2: change a condition "dr_mode"")
Signed-off-by: Fabrizio Castro <fabrizio.castro@bp.renesas.com>
Signed-off-by: Kishon Vijay Abraham I <kishon@ti.com>