arm64: dts: rockchip: increase the combphy refclk speed
The combphy currently is not completely known how it handles the clk
input, it can operate using it as a refclk for an internal pll or it can
pass it through directly.
Right now it seems to pass it through directly, so by upping the refclk
we can improve performance of the pcie and usb/sata controllers.

Signed-off-by: Peter Geis <pgwipeout@gmail.com>