arm64: dts: rockchip: fix rk3566-roc-pc pinctrl collisions
The pinctrl definitions were incorrect for i2c, causing a definition
collision.
Fix the incorrect definition, default the i2c devices on the
header to disabled.
Remove uart0, it can't exist anywhere.
Make the gpu regulator always on, the power domain driver can't handle
it being off yet.

This is round one of cleanup for this board.

Signed-off-by: Peter Geis <pgwipeout@gmail.com>